backend/requirements.txt
asgiref==3.8.1
gunicorn==23.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.4
djangorestframework-simplejwt
django-formtools==2.5.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
dotenv==0.9.9
idna==3.10
whitenoise==6.7.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.3.2
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.27.0
django-cors-headers==4.4.0
django-soft-delete==0.9.21
pre-commit===3.5.0
psycopg2-binary>=2.9
sendgrid==6.12.4
